backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.5
langchain-community>=0.2.0
duckduckgo-search>=5.2.2
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.6.4
